+++ src/hgvs/pretty/console/prot_seq_renderer.py
@@ -14,7 +14,7 @@ class ProtSeqRenderer(BasicRenderer):
         if not data.var_c_or_n:
             return ""
 
-        from hgvs.pretty.pretty_print import ENDC, TGREEN, TRED
+        from hgvs.pretty.console.constants import ENDC, TGREEN, TRED
 
         var_str = ""
         for pdata in data.position_details:
